backend/requirements.txt
fastapi
uvicorn
pymongo
ping3
psutil
requests
pydantic
dnspython